rename files with "lisp-case" naming
