+++ app/imports/schemas/galleryInformation.js
